+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Unreleased
 
 - Refresh the reading list near HH:00 The goal is to refresh the reading list closer to when the feeds where updated.
+- Hide the "Make default" button when creating a reading list.
+- Search improvements:
+  - Can search without a search text. This enables the usage of other fields without the need to find a text to search for.
+  - Can search for articles linked to specific feeds.
+  - Can search articles with some external tags.
+  - Can go to the advanced search page from all lists of articles pages. This allows users to start a search from a reading list, a feed page…
 
 ## 25.05.2